==> 5_Sentimental-score.ipynb <==
import pandas as pd

lexique_sent = pd.read_csv('data/externe/lexique_sentimental.csv')
train = pd.read_csv("./data/current/train.csv")
dev = pd.read_csv("./data/current/dev.csv")
test = pd.read_csv("./data/current/test.csv")

lex = lexique_sent['word'].values.tolist()
# Réaliser cela aussi pour dev et pour test
lexique = train['lemmatized'].apply(lambda x: [lexique_sent.loc[lexique_sent['word'] == i, 'polarity'].iloc[0] if i in lex else None for i in str(x).split()])
a = []
for i in lexique:
    if i is not None:
        t = 0
        x = 0
        for j in i:
            if j is not None:
                t += j
                x += 1
        if x != 0:
            a.append(t/x)
        else:
            a.append(None)
    else:
        a.append(None)
print(a)

train['lexique_sentimetal'] = a
train.to_csv('data/final/train.csv')
